fixed compiler warning (int -> NSInteger)

+++ Parsnip/MGInputDetailViewController.m
@@ -64,12 +64,12 @@
 
 #pragma mark Picker View
 
--(int)pickerView:(UIPickerView *)pickerView numberOfRowsInComponent:(NSInteger)component
+-(NSInteger)pickerView:(UIPickerView *)pickerView numberOfRowsInComponent:(NSInteger)component
 {
 	return 0;
 }
 
--(int)numberOfComponentsInPickerView:(UIPickerView *)pickerView
+-(NSInteger)numberOfComponentsInPickerView:(UIPickerView *)pickerView
 {
 	return 0;
 }
